Moves `ImageLoaderWithFallbackComposite` to production.
